Reorganize imports and remove extra function
